refactor: GoogleImageCloudService 클래스 이름 변경 및 기능개선
- GoogleImageCloudService 에서 GoogleBucketCloudService 변경
- CloudBucketService 인터페이스를 사용하여 강한 종속성을 제거
